Ответ на callback после вывода вакансий при пагинации. Изменения текстов

+++ vacancy-tracker-bot/app/src/main/java/vacancy_tracker/model/search/CollectedBatch.java
@@ -11,8 +11,8 @@ public class CollectedBatch {
 
     private final List<Vacancy> matched = new ArrayList<>();
 
-    private long nextPage;
+    private int nextPage;
     private long total;
-    private long nextOffset;
+    private int nextOffset;
     private boolean isSourceExhausted;
 }
